Public master underwent autoformatting. We should synchronize to avoid horrible merge conflicts later. There shouldn't be any (significant) changes apart from formatting compared to the previous merge.

Related work items: #90261, #90262

+++ src/data/corpus.h
@@ -57,7 +57,7 @@ public:
   std::vector<Ptr<Vocab>>& getVocabs() { return vocabs_; }
 
   batch_ptr toBatch(const std::vector<sample>& batchVector) {
-    int batchSize = batchVector.size();
+    size_t batchSize = batchVector.size();
 
     std::vector<size_t> sentenceIds;
 
@@ -73,14 +73,14 @@ public:
     }
 
     std::vector<Ptr<SubBatch>> subBatches;
-    for (int j = 0; j < maxDims.size(); ++j) {
+    for(size_t j = 0; j < maxDims.size(); ++j) {
       subBatches.emplace_back(New<SubBatch>(batchSize, maxDims[j], vocabs_[j]));
     }
 
     std::vector<size_t> words(maxDims.size(), 0);
-    for(int i = 0; i < batchSize; ++i) {
-      for(int j = 0; j < maxDims.size(); ++j) {
-        for(int k = 0; k < batchVector[i][j].size(); ++k) {
+    for(size_t i = 0; i < batchSize; ++i) {
+      for(size_t j = 0; j < maxDims.size(); ++j) {
+        for(size_t k = 0; k < batchVector[i][j].size(); ++k) {
           subBatches[j]->data()[k * batchSize + i] = batchVector[i][j][k];
           subBatches[j]->mask()[k * batchSize + i] = 1.f;
           words[j]++;
@@ -102,5 +102,5 @@ public:
     return batch;
   }
 };
-}
-}
+}  // namespace data
+}  // namespace marian
