+++ src/rnn/constructors.h
@@ -98,7 +98,7 @@ public:
 
     int lastDimInput = options_->get<int>("dimInput");
 
-    for(int i = 0; i < stackableFactories_.size(); ++i) {
+    for(size_t i = 0; i < stackableFactories_.size(); ++i) {
       auto sf = stackableFactories_[i];
 
       if(sf->is<CellFactory>()) {
@@ -142,7 +142,7 @@ public:
 
   Ptr<RNN> construct() {
     auto rnn = New<RNN>(graph_, options_);
-    for(int i = 0; i < layerFactories_.size(); ++i) {
+    for(size_t i = 0; i < layerFactories_.size(); ++i) {
       auto lf = layerFactories_[i];
 
       lf->getOptions()->merge(options_);
@@ -193,5 +193,5 @@ public:
 };
 
 typedef Accumulator<RNNFactory> rnn;
-}
-}
+}  // namespace rnn
+}  // namespace marian
