+++ src/tensors/cpu/element.h
@@ -17,24 +17,24 @@ namespace cpu {
 // on strides, it correctly broadcasts to all dimensions without additional
 // computation.
 // Compiler optimizes this to single construct with nested(?) loops.
-template <size_t I = 0> struct E {
+template <size_t I = 0>
+struct E {
   template <size_t K, class Functor>
-  static inline void element(const Functor& functor,
-                             functional::Array<functional::Tensor<float>, K>& tensors,
-                             functional::Array<int, K> indices) {
-
+  static inline void element(
+      const Functor& functor,
+      functional::Array<functional::Tensor<float>, K>& tensors,
+      functional::Array<int, K> indices) {
     const auto& shape = tensors[0].shape();
 
     // loop over outer-most dimension
     for(int i = 0; i < shape[I]; ++i) {
-
       // call loop for next-inner dimension
       E<I + 1>::element(functor, tensors, indices);
 
-      // increase index for current dimension by stride or 0 if broadcasting. bstride(i)
-      // is look-up value, either equal to stride if the corresponding dim is larger 1 or
-      // 0 if the dim is 1.
-      for(int k = 0; k < K; ++k)
+      // increase index for current dimension by stride or 0 if broadcasting.
+      // bstride(i) is look-up value, either equal to stride if the
+      // corresponding dim is larger 1 or 0 if the dim is 1.
+      for(size_t k = 0; k < K; ++k)
         indices[k] += tensors[k].shape().bstride(I);
     }
   }
@@ -42,15 +42,15 @@ template <size_t I = 0> struct E {
 
 // specialization for inner-most single element (recursive stopping criterion)
 // using const reference for indices here to avoid copying. No loop.
-template <> struct E<functional::Shape::size()> {
+template <>
+struct E<functional::Shape::size()> {
   template <size_t K, class Functor>
-  static inline void element(const Functor& functor,
-                             functional::Array<functional::Tensor<float>, K>& tensors,
-                             const functional::Array<int, K>& indices) {
-
+  static inline void element(
+      const Functor& functor,
+      functional::Array<functional::Tensor<float>, K>& tensors,
+      const functional::Array<int, K>& indices) {
     // just apply the function for all indexed elements across all tensors
     tensors[0][indices[0]] = functional::apply(functor, tensors, indices);
-
   }
 };
 
@@ -69,5 +69,5 @@ void Element(const Functor& functor, marian::Tensor out, Tensors... tensors) {
   E<0>::element(functor, gTensors, indices);
 }
 
-}
-}
+}  // namespace cpu
+}  // namespace marian
